+++ elasticsearch/src/main/java/org/apache/metamodel/elasticsearch/ElasticSearchDropTableBuilder.java
@@ -18,6 +18,8 @@
  */
 package org.apache.metamodel.elasticsearch;
 
+import java.lang.reflect.Method;
+
 import org.apache.metamodel.MetaModelException;
 import org.apache.metamodel.drop.AbstractTableDropBuilder;
 import org.apache.metamodel.drop.TableDropBuilder;
@@ -37,7 +39,7 @@ import org.slf4j.LoggerFactory;
 final class ElasticSearchDropTableBuilder extends AbstractTableDropBuilder {
 
     private static final Logger logger = LoggerFactory.getLogger(ElasticSearchDropTableBuilder.class);
-    
+
     private final ElasticSearchUpdateCallback _updateCallback;
 
     public ElasticSearchDropTableBuilder(ElasticSearchUpdateCallback updateCallback, Table table) {
@@ -56,11 +58,46 @@ final class ElasticSearchDropTableBuilder extends AbstractTableDropBuilder {
         final String indexName = dataContext.getIndexName();
 
         final DeleteMappingRequestBuilder requestBuilder = new DeleteMappingRequestBuilder(indicesAdminClient)
-                .setIndices(indexName).setType(documentType);
+                .setIndices(indexName);
+        setType(requestBuilder, documentType);
+
         final DeleteMappingResponse result = requestBuilder.execute().actionGet();
         logger.debug("Delete mapping response: acknowledged={}", result.isAcknowledged());
-        
+
         final MutableSchema schema = (MutableSchema) table.getSchema();
         schema.removeTable(table);
+    }
+
+    /**
+     * Invokes the {@link DeleteMappingRequestBuilder#setType(String...)} method
+     * using reflection. This is done because the API of ElasticSearch was
+     * changed and the method signature differes between different versions.
+     * 
+     * @param requestBuilder
+     * @param documentType
+     */
+    private void setType(DeleteMappingRequestBuilder requestBuilder, String documentType) {
+        Object argument;
+        Method method;
+        try {
+            try {
+                method = requestBuilder.getClass().getDeclaredMethod("setType", String[].class);
+                argument = new String[] {documentType};
+            } catch (NoSuchMethodException e) {
+                logger.debug("No setType(String[]) method found, trying with a single String instead", e);
+                method = requestBuilder.getClass().getDeclaredMethod("setType", String.class);
+                argument = documentType;
+            }
+        } catch (Exception e) {
+            logger.error("Failed to resolve DeleteMappingRequestBuilder.setType(...) method", e);
+            throw new IllegalStateException("Failed to resolve DeleteMappingRequestBuilder.setType(...) method", e);
+        }
+        try {
+            method.setAccessible(true);
+            method.invoke(requestBuilder, argument);
+        } catch (Exception e) {
+            logger.error("Failed to invoke {}", method, e);
+            throw new IllegalStateException("Failed to invoke " + method, e);
+        }
     }
 }